[v2] refactor `languages.py`
Refactored languages.py into a module. This should be fully backwards compatible.

Fixes #2808
