+++ mteb/languages/__init__.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from .check_language_code import check_language_code
+from .iso_mappings import ISO_TO_FAM, ISO_TO_FAM_LEVEL0, ISO_TO_LANGUAGE, ISO_TO_SCRIPT
+from .language_script import LanguageScripts
+from .programming_languages import PROGRAMMING_LANGS
+
+__all__ = [
+    "LanguageScripts",
+    "ISO_TO_LANGUAGE",
+    "ISO_TO_SCRIPT",
+    "ISO_TO_FAM",
+    "ISO_TO_FAM_LEVEL0",
+    "PROGRAMMING_LANGS",
+    "check_language_code",
+]
