+++ mteb/languages/language_script.py
@@ -1,51 +1,22 @@
-"""Language codes (ISO 639-3) obtained from: https://iso639-3.sil.org/sites/iso639-3/files/downloads/iso-639-3.tab
-Script codes (ISO 15924) obtained from: https://unicode.org/iso15924/iso15924.txt
-"""
-
 from __future__ import annotations
 
-import json
 from collections.abc import Iterable
 from dataclasses import dataclass
-from pathlib import Path
-
-# Language mappings
-path_to_lang_codes = Path(__file__).parent / "iso_639_3_to_language.json"
-path_to_lang_scripts = Path(__file__).parent / "iso_15924_to_script.json"
-path_to_lang_fam = Path(__file__).parent / "language_family.json"
-
-PROGRAMMING_LANGS = [
-    "python",
-    "javascript",
-    "typescript",
-    "go",
-    "ruby",
-    "java",
-    "php",
-    "c",
-    "c++",
-    "c#",
-    "rust",
-    "swift",
-    "scala",
-    "shell",
-    "sql",
-]
-
-with path_to_lang_codes.open("r") as f:
-    ISO_TO_LANGUAGE = json.load(f)
-
-with path_to_lang_scripts.open("r") as f:
-    ISO_TO_SCRIPT = json.load(f)
-
-with path_to_lang_fam.open("r") as f:
-    ISO_TO_FAM = json.load(f)
-
-ISO_TO_FAM_LEVEL0 = {k: v["level0"] for k, v in ISO_TO_FAM.items()}
 
 
 @dataclass
 class LanguageScripts:
+    """A utility class to represent a set of languages and scripts.
+
+    It can be used to check if a language or script is contained within the set.
+    It supports both ISO 639-3 language codes and ISO 15924 script codes.
+
+    Attributes:
+        language_scripts: A set of language-script combinations (e.g. "eng-Latn").
+        scripts: A set of script codes (e.g. "Latn").
+        languages: A set of language codes (e.g. "eng").
+    """
+
     language_scripts: set[str]
     scripts: set[str]
     languages: set[str]
@@ -77,6 +48,7 @@ class LanguageScripts:
         )
 
     def contains_language(self, language: str) -> bool:
+        """Whether the set contains a specific language."""
         if not self.language_scripts and not self.languages:
             return True
 
@@ -102,6 +74,7 @@ class LanguageScripts:
         return True
 
     def contains_script(self, script: str) -> bool:
+        """Whether the set contains a specific script."""
         return script in self.scripts
 
     def contains_scripts(self, scripts: Iterable[str]) -> bool:
@@ -110,23 +83,3 @@ class LanguageScripts:
             if not self.contains_script(s):
                 return False
         return True
-
-
-def check_language_code(code: str) -> None:
-    """This method checks that the language code (e.g. "eng-Latn") is valid."""
-    lang, script = code.split("-")
-    if script == "Code":
-        if lang in PROGRAMMING_LANGS:
-            return  # override for code
-        else:
-            raise ValueError(
-                f"Programming language {lang} is not a valid programming language."
-            )
-    if lang not in ISO_TO_LANGUAGE:
-        raise ValueError(
-            f"Invalid language code: {lang}, you can find valid ISO 639-3 codes in {path_to_lang_codes}"
-        )
-    if script not in ISO_TO_SCRIPT:
-        raise ValueError(
-            f"Invalid script code: {script}, you can find valid ISO 15924 codes in {path_to_lang_scripts}"
-        )
