+++ mteb/languages/programming_languages.py
@@ -1,0 +1,19 @@
+from __future__ import annotations
+
+PROGRAMMING_LANGS = [
+    "python",
+    "javascript",
+    "typescript",
+    "go",
+    "ruby",
+    "java",
+    "php",
+    "c",
+    "c++",
+    "c#",
+    "rust",
+    "swift",
+    "scala",
+    "shell",
+    "sql",
+]
